Add files via upload
DBUtil are utils I used in my projects to manage data and fill controls in VB6. With DBUtil I could generate SQL Queries giving just the tables and fields to get. This can be compiled in a DLL and, then, used in projects to get data from Access and SQL Server.